+++ koch.nim
@@ -11,7 +11,9 @@
 
 const
   NimbleStableCommit = "324de9202fb3db82b266e7350731d1ec41013a2b" # master
-
+  # examle of possible values: #head, #ea82b54, 1.2.3
+  FusionStableCommitHash = "#372ee4313827ef9f2ea388840f7d6b46c2b1b014"
+  HeadHash = "#head"
 when not defined(windows):
   const
     Z3StableCommit = "65de3f748a6812eecd7db7c478d5fc54424d368b" # the version of Z3 that DrNim uses
@@ -65,6 +67,7 @@ Possible Commands:
                            e.g. nimble)
                            doesn't require network connectivity
   nimble                   builds the Nimble tool
+  fusion                   installs fusion via Nimble
 
 Boot options:
   -d:release               produce a release version of the compiler
@@ -689,7 +692,9 @@ when isMainModule:
       of "valgrind": valgrind(op.cmdLineRest)
       of "c2nim": bundleC2nim(op.cmdLineRest)
       of "drnim": buildDrNim(op.cmdLineRest)
-      of "fusion": doAssert false, "`./koch fusion` was deprecated, use `nimble install fusion` instead"
+      of "fusion":
+        let suffix = if latest: HeadHash else: FusionStableCommitHash
+        exec("nimble install -y fusion@$#" % suffix)
       else: showHelp()
       break
     of cmdEnd: break