+++ changelog.md
@@ -188,5 +188,5 @@ with other backends. see #9125. Use `-d:nimLegacyJsRound` for previous behavior.
     row (`:---`, `:--:`, `---:`) are ignored,
   - every table row must start with `|`, e.g. `| cell 1 | cell 2 |`.
 
-- `fusion` is now un-bundled from nim, `./koch fusion` will result in an error.
-  `nimble install fusion` should be used instead.
+- `fusion` is now un-bundled from nim, `./koch fusion` will
+  install it via nimble at a fixed hash.